+++ src/ui/components/bottomSheet/index.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { BottomSheetModal, BottomSheetBackdrop, BottomSheetBackdropProps, BottomSheetModalProvider, BottomSheetView } from '@gorhom/bottom-sheet';
+
+// Re-export components from @gorhom/bottom-sheet
+// This allows us to use these components internally without requiring the consumer to install @gorhom/bottom-sheet
+export {
+    BottomSheetModal,
+    BottomSheetBackdrop,
+    BottomSheetBackdropProps,
+    BottomSheetModalProvider,
+    BottomSheetView
+};